requirements.txt
Flask==2.2.3
flask-cors==3.0.10
transformers==4.30.2
torch==2.0.1
youtube-transcript-api==0.6.0
python-docx==0.8.11
